+++ src/webapp/src/components/analyze/AnalyzeList.vue
@@ -112,8 +112,8 @@ export default {
         <div class="buttons">
           <router-link
             v-for="design in model['designs']"
-            tag="button"
             :key="design"
+            tag="button"
             class="button is-small is-interactive-primary is-outlined"
             :disabled="!getIsModelEnabled(model)"
             :to="urlForModelDesign(modelKey, design)"